Refactor TypeScript interfaces for consistency
